ENH: Handles images adresses in both SAFE and SAFE-COMPACT case
Image naming in SAFE-COMPACT mode requires an Id that can only be read from the main MTD file. As a results, several hacks are necessary to be able to read granule MTD and access SENTINEL2_L1C_TILE: subdatasets:

- GetGranuleList() method also retrieves the image base name depending on format
- CreateL1CL2ADataset() needs as extra input the list of those image base name
- SENTINEL2GetMainMTDFilenameFromGranuleMTD() needed to be updated to support SAFE-COMPACT format
- OpenL1CTileSubdataset() needs to re-open main MTD file to actually know how image names are formed.

This means that tile subdatasets can not be read if the main MTD can not be found.